Remove link underlines; indicate state by colour
Except for tags sidebar and page footer links.

*

There don't seem to be any good solutions for the case where there are
links which have text containing <code> and non-code text.

Part of the problem is the padding on the code which looks better than
not having it, but means that there is a gap in the underlining.

GitHub's formatting is:

code, tt {
    padding: 0.2em 0px;
    margin: 0px;
    font-size: 85%;
    background-color: rgba(0, 0, 0, 0.04);
    border-radius: 3px;
}

code::before, code::after {
    letter-spacing: -0.2em;
    content: "\00a0";
}

This looks OK in Firefox, though has a gap inside the spacing with no
underline. However, it looks bad in Chromium.

Stack Overflow doesn't underline links at all.

In the end, it seemed simpler to just remove underlines completely.