Use font-size-base of 17px instead of 22px
Bootstrap 3 is mobile-first. So, use mobile-friendly font sizes and
then override them for larger displays.

Overrides obtained by diffing the Bootswatch for 17px and 22px
font-size-base.

One issue is that there may be property values that are overridden later
in the Bootswatch. The differences added here could arise from these types
of cases. The only one found to be detrimental to the appearance on large
screens was the commented out navbar-nav margin.

(This isn't an ideal solution, by any means.)

Line heights decided by finding the closest px integer value that gave
a value close to the Bootstrap default of 20/14 for the font-size-base.