Update README.md
Cherry picked from commit
DandyDev/pelican-bootstrap3@e4312577fc60727f860def19d31e33332cad1751.

This is the last relevant commit from the DandyDev/pelican-bootstrap3
repository before it was moved to getpelican/pelican-themes. The commit
following this deleted the content and edited the README.md to point the
reader to the pelican-themes repository.