pelican-bootstrap3: refactored sidebar (#466)
This is more of a proposal for refactor of the sidebar.  Figured it
would be easier to do and discuss in a PR than it would be an
issue/ticket.

These changes can be split out if only some of the ideas are accepted.
It is mostly a lot cleanup and shuffling around to make it a bit more
consistent.

Here is a summary of the changes:

  - Moved all sidebar related from `includes/` and into
    `includes/sidebar/`.
  - Moved the following sections in `includes/sidebar.html` into
    their own files under `includes/sidebar/`:
    - _Social_ -> `includes/sidebar/social.html`
    - _Recent Posts_ -> `includes/sidebar/recent_posts.html`
    - _Categories_ -> `includes/sidebar/categories.html`
    - _Tags_ -> `includes/sidebar/tag_cloud.html`
    - _Series_ -> `includes/sidebar/series.html`
  - A whole bunch of whitespace cleanup.
  - Added the `DISABLE_SIDEBAR_TITLE_ICONS` configuration option. If
    `DISABLE_SIDEBAR_TITLE_ICONS` is set to True, none of the sidebar
    title icons will be displayed.
  - Folded in from getpelican/pelican-themes#438: Added the
    `SIDEBAR_ON_LEFT` configuration option, which moves the sidebar
    to the left side of the page.  Sidebar will continue to appear on
    the bottom of the page on smaller screens.
  - Assuming getpelican/pelican#1092 worked out:
    - Optional include of `includes/sidebar/optional_top.html` and
      `includes/sidebar/optional_bottom.html` from within
      `includes/sidebar.html`.  This would allow a user to insert
      sections into the sidebar.
    - `includes/sidebar/recent_posts.html` include articles line items
      via `includes/sidebar/article-li.html`.  This allows customization.

Cherry picked from commit
getpelican/pelican-themes@ae7a0db55340f646f9eb3d15d323959040ece7e8.

Conflict due to pelican-bootstrap3-sm having the DISABLE_TRANSITIONS
option, but didn't have any net impact on this commit.