Move into pelican-bootstrap3 directory
The getpelican/pelican-themes version of pelican-bootstrap3 has
everything in a pelican-bootstrap3 directory. Without this, some `git
cherry-pick` operations get more complicated, because, for some reason,
git can't resolve the differences between the source repository and this
fork where the commit is to be applied.

The simplest fix is to reorganise this repository. It does have the
consequence that existing sites using this forked theme will have to
change their configuration or move the files manually to the correct
location that they are required.